requirements.txt
# Core dependencies
python-dotenv==1.0.0

# LLM and Embeddings
openai==1.12.0

# Vector Database
chromadb==0.4.22

# MCP Server
flask==3.0.0

# Database
oracledb==2.0.0
sqlalchemy==2.0.25

# Utilities
requests==2.31.0
